Removed the cyclical dependency in workspaces and announcement (#1519)
