+++ src/features/announcements/constants.js
@@ -1,0 +1,5 @@
+export const ANNOUNCEMENTS_ROUTES = {
+  TARGET: '/announcements/:id',
+};
+
+export const GA_CATEGORY_ANNOUNCEMENTS = 'Announcements';
