+++ src/features/announcements/index.js
@@ -7,9 +7,6 @@ export const GA_CATEGORY_ANNOUNCEMENTS = 'Announcements';
 
 export const announcementsStore = new AnnouncementsStore();
 
-export const ANNOUNCEMENTS_ROUTES = {
-  TARGET: '/announcements/:id',
-};
 
 export default function initAnnouncements(stores, actions) {
   const { features } = stores;
