+++ src/features/workspaces/constants.js
@@ -1,0 +1,7 @@
+export const WORKSPACES_ROUTES = {
+  ROOT: '/settings/workspaces',
+  EDIT: '/settings/workspaces/:action/:id',
+};
+
+export const GA_CATEGORY_WORKSPACES = 'Workspaces';
+export const DEFAULT_SETTING_KEEP_ALL_WORKSPACES_LOADED = false;
